telegram-booking-bot/backend/bot.py
import os
import asyncio
import logging

from dotenv import load_dotenv
from telegram import Update, WebAppInfo, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import Application, CommandHandler, ContextTypes

load_dotenv(os.path.join(os.path.dirname(__file__), "..", ".env"))

BOT_TOKEN = os.getenv("BOT_TOKEN", "")
WEBAPP_URL = os.getenv("WEBAPP_URL", "")

logging.basicConfig(
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
    level=logging.INFO,
)
logger = logging.getLogger(__name__)


async def start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """Обработчик команды /start."""
    keyboard = [
        [
            InlineKeyboardButton(
                text="\U0001f485 Записаться",
                web_app=WebAppInfo(url=WEBAPP_URL),
            )
        ]
    ]
    reply_markup = InlineKeyboardMarkup(keyboard)

    await update.message.reply_text(
        text=(
            "\U0001f44b Добро пожаловать в салон красоты!\n\n"
            "Нажмите кнопку ниже, чтобы записаться на услугу:"
        ),
        reply_markup=reply_markup,
    )


def main() -> None:
    if not BOT_TOKEN:
        logger.error("BOT_TOKEN is not set!")
        return
    if not WEBAPP_URL:
        logger.error("WEBAPP_URL is not set!")
        return

    application = Application.builder().token(BOT_TOKEN).build()
    application.add_handler(CommandHandler("start", start))

    logger.info("Bot started")
    application.run_polling(allowed_updates=Update.ALL_TYPES)


if __name__ == "__main__":
    main()
